refactor: MenuDetailContent와 NavBar의 의존성을 업데이트하고, FailPage 컴포넌트를 단순화
- MenuDetailContent의 useEffect에 holdTimeout을 의존성 배열에 추가
- NavBar의 useEffect에 setIsFixed를 의존성 배열에 추가
- FailPage를 Suspense와 함께 FailClient 컴포넌트를 사용
